requirements.txt
Flask
python-dotenv
pymongo
playwright
aiohttp
